requirements.txt
Django==4.2.7
gunicorn==21.2.0
whitenoise==6.6.0

djangorestframework==3.14.0
django-widget-tweaks==1.5.1
django-cors-headers==4.9.0
django-allauth==65.15.0

python-decouple==3.8
python-dotenv==1.0.0

requests==2.31.0
google-genai==0.3.0